refactor form list
