+++ packages/react-editor/src/website/pages/FormList/header-actions.js
@@ -1,0 +1,8 @@
+export default ({ resetDB, create }) => [{
+  label: 'Init Mock Forms',
+  onClick: resetDB,
+  variant: 'outlined',
+}, {
+  label: 'Create',
+  onClick: create,
+}];
